Extract FieldInferenceIngestService
