+++ server/src/main/java/org/elasticsearch/ingest/FieldInferenceIngestService.java
@@ -1,0 +1,128 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+package org.elasticsearch.ingest;
+
+import org.elasticsearch.action.ActionListener;
+import org.elasticsearch.action.DocWriteRequest;
+import org.elasticsearch.action.bulk.TransportBulkAction;
+import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.action.inference.InferenceAction;
+import org.elasticsearch.action.support.RefCountingRunnable;
+import org.elasticsearch.client.internal.Client;
+import org.elasticsearch.cluster.service.ClusterService;
+import org.elasticsearch.common.util.concurrent.AbstractRunnable;
+import org.elasticsearch.env.Environment;
+import org.elasticsearch.grok.MatcherWatchdog;
+import org.elasticsearch.index.analysis.AnalysisRegistry;
+import org.elasticsearch.inference.TaskType;
+import org.elasticsearch.plugins.IngestPlugin;
+import org.elasticsearch.plugins.internal.DocumentParsingObserver;
+import org.elasticsearch.script.ScriptService;
+import org.elasticsearch.threadpool.ThreadPool;
+
+import java.util.List;
+import java.util.Map;
+import java.util.function.BiConsumer;
+import java.util.function.IntConsumer;
+import java.util.function.Supplier;
+
+public class FieldInferenceIngestService extends IngestService {
+
+    public FieldInferenceIngestService(ClusterService clusterService, ThreadPool threadPool, Environment env, ScriptService scriptService, AnalysisRegistry analysisRegistry, List<IngestPlugin> ingestPlugins, Client client, MatcherWatchdog matcherWatchdog, Supplier<DocumentParsingObserver> documentParsingObserverSupplier) {
+        super(clusterService, threadPool, env, scriptService, analysisRegistry, ingestPlugins, client, matcherWatchdog, documentParsingObserverSupplier);
+    }
+
+    public void executeFieldInferenceBulkRequest(
+        final int numberOfActionRequests,
+        final Iterable<DocWriteRequest<?>> actionRequests,
+        final IntConsumer onDropped,
+        final BiConsumer<Integer, Exception> onFailure,
+        final BiConsumer<Thread, Exception> onCompletion,
+        final String executorName
+    ) {
+        assert numberOfActionRequests > 0 : "numberOfActionRequests must be greater than 0 but was [" + numberOfActionRequests + "]";
+
+        threadPool.executor(executorName).execute(new AbstractRunnable() {
+
+            @Override
+            public void onFailure(Exception e) {
+                onCompletion.accept(null, e);
+            }
+
+            @Override
+            protected void doRun() {
+                final Thread originalThread = Thread.currentThread();
+                try (var refs = new RefCountingRunnable(() -> onCompletion.accept(originalThread, null))) {
+                    int i = 0;
+                    for (DocWriteRequest<?> actionRequest : actionRequests) {
+                        IndexRequest indexRequest = TransportBulkAction.getIndexWriteRequest(actionRequest);
+                        if (indexRequest != null) {
+                            String index = indexRequest.index();
+                            Map<String, Object> sourceMap = indexRequest.sourceAsMap();
+                            final int position = i;
+                            sourceMap.entrySet().stream().filter(entry -> fieldNeedsInference(index, entry.getKey())).forEach(entry -> {
+                                runInferenceForField(indexRequest, entry.getKey(), entry.getValue(), refs, position, onFailure);
+                            });
+                        }
+                        i++;
+                    }
+                }
+            }
+        });
+    }
+    public boolean needsFieldInference(IndexRequest indexRequest) {
+        return (indexRequest.isFieldInferenceResolved() == false)
+            && indexRequest.sourceAsMap().keySet().stream().anyMatch(fieldName -> fieldNeedsInference(indexRequest.index(), fieldName));
+    }
+
+    // TODO actual mapping check here
+    private boolean fieldNeedsInference(String index, String fieldName) {
+        return fieldName.startsWith("infer_");
+    }
+
+    private void runInferenceForField(
+        IndexRequest indexRequest,
+        String fieldName,
+        Object fieldValue,
+        RefCountingRunnable ref,
+        int position,
+        BiConsumer<Integer, Exception> onFailure
+    ) {
+        var ingestDocument = newIngestDocument(indexRequest, documentParsingObserverSupplier.get());
+        if (ingestDocument.hasField(fieldName) == false) {
+            return;
+        }
+
+        ref.acquire();
+
+        // TODO Hardcoding model ID and task type
+        InferenceAction.Request inferenceRequest = new InferenceAction.Request(
+            TaskType.SPARSE_EMBEDDING,
+            "my-elser-model",
+            ingestDocument.getFieldValue(fieldName, String.class),
+            Map.of()
+        );
+
+        client.execute(InferenceAction.INSTANCE, inferenceRequest, new ActionListener<InferenceAction.Response>() {
+            @Override
+            public void onResponse(InferenceAction.Response response) {
+                ingestDocument.setFieldValue(fieldName + "_inference", response.getResult().asMap(fieldName).get(fieldName));
+                updateIndexRequestSource(indexRequest, ingestDocument);
+                indexRequest.isFieldInferenceResolved(true);
+                ref.close();
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                onFailure.accept(position, e);
+                ref.close();
+            }
+        });
+    }
+}
